New spool, new mess
Changes from yesterday's stream to add spool printability relief holes
and a new knot recess in the center hub. These ended up making a huge
mess of the already-completed frame component, so I've attempted to fix
that up by reverting to an older version of the F3D file and
re-importing the latest spool from STEP. This has reset the F360 version
numbering.